Smart Download functionality
Added a cvar with which the server owner can specify whether the client will receive all referencedpaks or just the one which contains the current map being played.

This would ensure a faster connect time to the server, combining that with peak 1MB/s download rate of SoF2Plus, this will make most map switches even when you don't have a map, instantaneous.