Fixed autodownloads when sv_clientmod is used
Because we spoof fs_game with sv_clientMod, there's an issue with autodownloader. Autodownloader will try to download from the actual fs_game folder, which will never be referenced by the client game because we send a spoofed fs_game to them.

So the workaround is to spoof the referenced paks names the client sees to the sv_clientMod value as well, and when a download request comes in with the sv_clientMod being the folder, we undo the spoof and feed the file from the correct directory.