Fix clipping issue on jumps + uncrouch
In SoF2 disassembly, there was a difference in the position test where there was an additional check on the half extents of the boundingbox. This seemed to also be done in Jedi Academy, but not in Jedi Outcast and Quake 3.

This fixes the clipping issue when you're doing a jump and then releasing crouch at the apex of your jump.