Increase MAX_SUBMODELS and MODEL_HANDLE's.
This fixes a crash when a heavily modified SP map (e.g. hk4) is loaded on the engine. HK4 has a lot of its own BSP's loaded on top of the entity file itself. Whilst most likely there is a workaround through 1fxmod for it (either by not spawning some instances or similar), having a higher set of model handles so they do not interfere with actual map models solves it.

For the HK4 case, a BOX_MODEL_HANDLE got linked to a tree BSP file, which caused more issues further downstream, as a box model should be a "dummy" model.