Respect multiprotocol state on server packet transmission.
The engine was built in a way that if net_multiprotocol is not set, then it should function as a SoF2 Gold engine. But this did not seem to function because packets were not transmitted due to legacy_ip_socket pointing to INVALID_SOCKET.

The check is now done on a packet-basis. legacyProtocol will never be set if we're in a non-multiprotocol game.