Limit getstatus response size, remove a couple of cvars from serverinfo
When a server heartbeats the official masterserver, it responds with a getstatus query. Current working theory is, that if the getstatus query packet gets fragmented, then the server info is ignored.

com_gamename was removed as it was redundant.
legacyprotocol was removed from serverinfo because it didn't have a purpose in serverinfo.